Refactor stories structures
