+++ main/index.ts
@@ -1,15 +1,8 @@
-export {Accordion} from './Accordion';
-export {Button} from './Button';
-export {ButtonGroup} from './ButtonGroup';
-export {EditText} from './EditText';
-export {LoadingIndicator} from './LoadingIndicator';
-export {StatusBarBrightness} from './StatusBarBrightness';
-export {PinchZoom, PinchZoomRef} from './PinchZoom';
-export {
-  ThemeProvider,
-  useTheme,
-  withTheme,
-  theme,
-  ThemeType,
-  DoobooTheme,
-} from './theme';
+export * from './Button';
+export * from './ButtonGroup';
+export * from './EditText';
+export * from './LoadingIndicator';
+export * from './Accordion';
+export * from './StatusBarBrightness';
+export * from './PinchZoom';
+export * from './theme';
