--- packages/TinderCard/. npmignore
@@ -1,3 +1,0 @@
-tsconfig.tsbuildinfo
-index.tsx
-tsconfig.json
