--- packages/TinderCard/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@dooboo-ui/tinder-card",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "version": "0.0.2",
-  "license": "MIT",
-  "postinstall": "tsc",
-  "scripts": {
-    "build": "tsc",
-    "watch": "tsc -w"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/dooboolab/dooboo-ui.git",
-    "directory": "src/components/shared/TinderCard"
-  },
-  "bugs": {
-    "url": "https://github.com/dooboolab/dooboo-ui/issues"
-  },
-  "homepage": "https://github.com/dooboolab/dooboo-ui#readme",
-  "dependencies": {
-    "dooboolab-welcome": "^1.2.0"
-  },
-  "devDependencies": {
-    "typescript": "^4.2.4"
-  },
-  "peerDependencies": {
-    "react": ">=16.9",
-    "react-native": ">=0.58",
-    "@emotion/native": "^11.0.0",
-    "@emotion/react": "^11.1.5"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "keywords": [
-    "swtich-toggle",
-    "react-hook",
-    "react-native",
-    "dooboo-ui/native",
-    "dooboolab"
-  ]
-}
