--- packages/TinderCard/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfigPackage.json",
-  "compilerOptions": {
-    "outDir": "./lib"
-  },
-  "exclude": ["**/*.d.ts", "node_modules"]
-}
