+++ stories/packages/PinzoomStories/PinchZoomSlider.tsx
